Fix tests and comments
